+++ flat.h
@@ -226,6 +226,6 @@ static inline uint32_t GetMsTicks(void) {
 
 cPixmap *CreatePixmap(cOsd *osd, int Layer = 0, const cRect &ViewPort = cRect::Null, const cRect &DrawPort = cRect::Null);
 cPlugin *GetScraperPlugin(void);
-cString GetSimpleAspectIcon(int screenWidth, float screenAspect);
-cString GetScreenResolutionIcon(int screenWidth, int screenHeight, float screenAspect);
+cString GetSimpleAspectIcon(int screenWidth, double screenAspect);
+cString GetScreenResolutionIcon(int screenWidth, int screenHeight, double screenAspect);
 cString GetFormatIcon(int screenWidth);
